cast x to double

+++ table.c
@@ -59,9 +59,9 @@ void set_table_matrix(struct table *, int);
 static double
 weight3(int x)
 {
-    if (x < 0.1)
+    if ((double)x < 0.1)
 	return 0.1;
-    if (x < LOG_MIN)
+    if ((double)x < LOG_MIN)
 	return (double)x;
     else
 	return LOG_MIN * (log((double)x / LOG_MIN) + 1.);
